Fix the request_data bug
I fixed the bug that followed the rename of the particular method. The
'request_data' method has been renamed to 'get_data' creating this bug
(because the code isn't well tested either). Now everything is in order
and I changed the variable names in the CLI to keep a sort of coherence
in the naming.

Resolves: #17

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>